+++ packages/@guardian/source-foundations/src/space.test.ts
@@ -1,4 +1,4 @@
-import { space, remSpace } from './space';
+import { remSpace, space } from './space';
 
 it('should provide a rem equivalent of the pixel space', () =>
 	expect(remSpace[2]).toBe(`${space[2] / 16}rem`));
