+++ packages/@guardian/source-foundations/src/utils/px-to-rem.ts
@@ -1,0 +1,6 @@
+// FYI
+// src/core/foundations/src/utils/px-to-rem.ts SYMLINKS TO
+// packages/@guardian/source-foundations/src/utils/px-to-rem.ts
+
+export const rootPixelFontSize = 16;
+export const pxToRem = (px: number): number => px / rootPixelFontSize;
