+++ src/core/foundations/src/theme.ts
@@ -38,8 +38,6 @@ const lineHeights = [1.15, 1.35, 1.5];
 
 const fontWeights = [300, 400, 500, 700];
 
-const space = [0, 4, 8, 12, 16, 20, 24, 36, 48, 96];
-
 const size = [24, 36, 44];
 
-export { fontSizes, fonts, lineHeights, fontWeights, space, size };
+export { fontSizes, fonts, lineHeights, fontWeights, size };
