+++ src/core/foundations/src/utils/px-to-rem.ts
@@ -1,2 +1,1 @@
-export const rootPixelFontSize = 16;
-export const pxToRem = (px: number): number => px / rootPixelFontSize;
+../../../../../packages/@guardian/source-foundations/src/utils/px-to-rem.ts
